requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
colorama==0.3.9
diff-match-patch==20121119
Django==2.0.6
django-import-export==1.0.1
django-redis==4.9.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
jdcal==1.4
more-itertools==4.2.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.4
pluggy==0.6.0
py==1.5.3
pytest==3.6.1
pytest-django==3.3.0
pytz==2018.4
PyYAML==3.12
redis==2.10.6
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
